move pytest entrypoint from __init__ into its own module
